face-emotion/requirements.txt
torch
torchvision
albumentations
opencv-python
numpy
pandas
scikit-learn
matplotlib
tqdm
fastapi
uvicorn[standard]
Pillow
pytest
